+++ src/BufferStream.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace ImpreseeGuzzleHttp\Psr7;
 
-use Psr\Http\Message\StreamInterface;
+use Impresee\Psr\Http\Message\StreamInterface;
 
 /**
  * Provides a buffer stream that can be written to to fill a buffer, and read