+++ src/MessageTrait.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace ImpreseeGuzzleHttp\Psr7;
 
-use Psr\Http\Message\MessageInterface;
-use Psr\Http\Message\StreamInterface;
+use Impresee\Psr\Http\Message\MessageInterface;
+use Impresee\Psr\Http\Message\StreamInterface;
 
 /**
  * Trait implementing functionality common to requests and responses.