+++ src/Request.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace ImpreseeGuzzleHttp\Psr7;
 
 use InvalidArgumentException;
-use Psr\Http\Message\RequestInterface;
-use Psr\Http\Message\StreamInterface;
-use Psr\Http\Message\UriInterface;
+use Impresee\Psr\Http\Message\RequestInterface;
+use Impresee\Psr\Http\Message\StreamInterface;
+use Impresee\Psr\Http\Message\UriInterface;
 
 /**
  * PSR-7 request implementation.