+++ src/Uri.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace ImpreseeGuzzleHttp\Psr7;
 
 use ImpreseeGuzzleHttp\Psr7\Exception\MalformedUriException;
-use Psr\Http\Message\UriInterface;
+use Impresee\Psr\Http\Message\UriInterface;
 
 /**
  * PSR-7 URI implementation.
@@ -149,7 +149,7 @@ class Uri implements UriInterface, \JsonSerializable
      * Composes a URI reference string from its various components.
      *
      * Usually this method does not need to be called manually but instead is used indirectly via
-     * `Psr\Http\Message\UriInterface::__toString`.
+     * `Impresee\Psr\Http\Message\UriInterface::__toString`.
      *
      * PSR-7 UriInterface treats an empty component the same as a missing component as
      * getQuery(), getFragment() etc. always return a string. This explains the slight
@@ -196,7 +196,7 @@ class Uri implements UriInterface, \JsonSerializable
     /**
      * Whether the URI has the default port of the current scheme.
      *
-     * `Psr\Http\Message\UriInterface::getPort` may return null or the standard port. This method can be used
+     * `Impresee\Psr\Http\Message\UriInterface::getPort` may return null or the standard port. This method can be used
      * independently of the implementation.
      */
     public static function isDefaultPort(UriInterface $uri): bool