+++ tests/NoSeekStreamTest.php
@@ -6,7 +6,7 @@ namespace ImpreseeGuzzleHttp\Tests\Psr7;
 
 use ImpreseeGuzzleHttp\Psr7\NoSeekStream;
 use PHPUnit\Framework\TestCase;
-use Psr\Http\Message\StreamInterface;
+use Impresee\Psr\Http\Message\StreamInterface;
 
 /**
  * @covers ImpreseeGuzzleHttp\Psr7\NoSeekStream