+++ tests/RequestTest.php
@@ -8,7 +8,7 @@ use ImpreseeGuzzleHttp\Psr7;
 use ImpreseeGuzzleHttp\Psr7\Request;
 use ImpreseeGuzzleHttp\Psr7\Uri;
 use PHPUnit\Framework\TestCase;
-use Psr\Http\Message\StreamInterface;
+use Impresee\Psr\Http\Message\StreamInterface;
 
 /**
  * @covers ImpreseeGuzzleHttp\Psr7\MessageTrait