+++ tests/StreamDecoratorTraitTest.php
@@ -7,7 +7,7 @@ namespace ImpreseeGuzzleHttp\Tests\Psr7;
 use ImpreseeGuzzleHttp\Psr7;
 use ImpreseeGuzzleHttp\Psr7\StreamDecoratorTrait;
 use PHPUnit\Framework\TestCase;
-use Psr\Http\Message\StreamInterface;
+use Impresee\Psr\Http\Message\StreamInterface;
 
 class Str implements StreamInterface
 {